Refactor to remove dumb friendship package, and lots of other changes. Everything is pretty.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -1,9 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+    use Illuminate\Database\Migrations\Migration;
+    use Illuminate\Database\Schema\Blueprint;
 
-class CreateUsersTable extends Migration
+    class CreateUsersTable extends Migration
 {
     /**
      * Run the migrations.
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password', 60);
-            $table->double('balance', 12, 2)->default(0);
+            $table->decimal('balance', 8, 2)->default(0);
             $table->string('key_id')->nullable();
             $table->timestamp('last_seen')->nullable();
             $table->string('picture_id')->nullable();
